Replace old SegmentControl fix #1874
